+++ packages/exr/src/encode.ts
@@ -1,4 +1,5 @@
 import {
+  importModule,
   isMultiThreadSupported,
   validateThreadCount,
   copyToWasm16f,
@@ -48,7 +49,7 @@ export async function init({
     const moduleConfig: Record<string, unknown> = {
       mainScriptUrlOrBlob: isMultiThreadedModule ? url : undefined,
     };
-    const module: WasmModule = await import(/* @vite-ignore */ url);
+    const module = await importModule<WasmModule>(url);
     const createModule = module.default;
     encoderModule = await createModule(moduleConfig);
     maxThreads = encoderModule.MAX_THREADS ?? 1;